feat(appium): cull invalid data from manifest
If, for some reason, an extension doesn't exist on disk yet does exist in the manifest, it probably shouldn't be in the manifest, right? Anyway:

This change causes invalid entries to be culled in the same operation as "sync with installed packages". Those situations are:

- if the manifest file is new
- if the manifest needs a migration
- if `appium` is a local dep and its `package.json` has changed

If none of these are true, synchronization won't happen (extensions won't be automatically discovered), and entries won't be culled.

An argument could be made that we want to sync and/or cull invalid entries regardless of whether we're using a local `appium` or `APPIUM_HOME`, and likewise on any extension subcommand which can modify the manifest (install, uninstall, update). While good housekeeping is in order, it's not free--the system must inspect _every `package.json` in `node_modules`_ to find extensions.

In fact, I'm thinking this should be _opt-in_ behavior if `appium` is a locally-installed dependency; the sync would happen any time `package.json` is modified on-disk.  This might be something like a flag `--sync` flag for `appium driver/plugin list`.  Anyway: that can be a separate PR.